changes per greg: clearer implementation, organizational fixes
