+++ templates/event_details.html
@@ -78,7 +78,7 @@
 	
 	<ul class="nav nav-tabs nav-stacked-phone">
 	  <li class="active"><a href="#results" data-toggle="tab">Results</a></li>
-      {% if rankings %}<li><a href="#rankings" data-toggle="tab">Rankings</a></li>{% endif %}
+      {% if event.rankings %}<li><a href="#rankings" data-toggle="tab">Rankings</a></li>{% endif %}
       <li><a href="#teams" data-toggle="tab">Teams</a></li>
 	  <!-- <li><a href="#awards" data-toggle="tab">Awards</a></li>-->
 	</ul>
@@ -152,12 +152,14 @@
 	  </div>
     </div>
 
-    {% if rankings %}
+    {% if event.rankings %}
     <div class="tab-pane" id="rankings">
         <div class="row">
             <div class="span8 offset2">
+            <p>Data from <a href="http://www2.usfirst.org/{{event.year}}comp/events/{{event.event_short}}/rankings.html" target="_blank">
+                <em>FIRST</em>'s event page</a>.</p>
             <table class="table table-striped table-condensed table-center">
-	        {% for row in rankings %}
+	        {% for row in event.rankings %}
 	            <tr>
 	            {% if forloop.first %}
                     {% for element in row %}
